+++ src/pages/Community/types/PostData.ts
@@ -1,14 +1,3 @@
-// export interface Post {
-//   id: number;
-//   title: string;
-//   author: string;
-//   timestamp: string;
-//   views: number;
-//   comments: number;
-//   // likes: number;
-//   content?: string; // content는 PostDetails에서만 사용되므로 optional로 지정
-// }
-
 export interface Post {
   data: {
     boardId: number; // 게시판 구분
